Added in the RegisterViewModel a test for the input email for when there is already an existing account with the same credentials, a toast message will pop up. For register, the success case leads now to the login screen, and in case of failure, the text boxes for email and password will be cleared.
Added in the registerViewModel a test for the input email for when there
is an already existing account with the same credentials, a toats
message will pop up. For register, the success case leads now to the
login screen, and in case of failure, the text boxes for email and
password will be cleared.